backend/requirements.txt
accelerate==1.10.1
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==4.0.3
attrs==25.3.0
awscli==1.42.35
awscurl==0.36
boto3==1.40.35
botocore==1.40.35
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
ConfigArgParse==1.7.1
configparser==7.2.0
docutils==0.19
fastapi==0.117.1
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
gremlinpython==3.7.4
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.35.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.6.4
nest-asyncio==1.6.0
networkx==3.3
numpy==2.3.3
packaging==25.0
pillow==11.0.0
propcache==0.3.2
psutil==7.1.0
pyasn1==0.6.1
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
requests-aws4auth==1.3.1
rsa==4.7.2
s3transfer==0.14.0
safetensors==0.6.2
setuptools==70.2.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.13.3
tokenizers==0.22.1
torch
torchaudio
torchvision
tqdm==4.67.1
transformers==4.56.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.36.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
